simplify conda installation

+++ scripts/setup_conda_env.sh
@@ -20,53 +20,36 @@ if [ -z "${SL3A_CONDA_ROOT}" ]; then
     exit 255
 fi
 
+if [ -f "${SL3A_CONDA_ROOT}/.marker.env.completed" ]; then
+    echo "Environment already installed"
+    exit 0
+fi
+
 SL3A_ENV_NAME="${SL3A_ENV_NAME:-sl3a}"
-SL3A_ENV_NAME_TMP=tmp
 
 CONDA="${SL3A_CONDA_ROOT}/miniconda3/bin/conda"
 CONDA_ACTIVATE="${SL3A_CONDA_ROOT}/miniconda3/bin/activate"
 
-if [ ! -z "${SL3A_CONDA_PERFORM_SETUP}" ]; then
-    if [ ! -f "${CONDA}" ]; then
-        echo "\"conda\" not found, check setup_conda_bin.sh"
-        exit 255
-    fi
-
-    if "${CONDA}" env list | grep -qw "^${SL3A_ENV_NAME}"; then
-        echo "Conda environment \"${SL3A_ENV_NAME}\" already exists"
-        exit 0
-    fi
-
-    "${CONDA}" install -y -n base --solver classic conda-libmamba-solver
-    "${CONDA}" config --set solver libmamba
-
-    "${CONDA}" create -y -n ${SL3A_ENV_NAME_TMP} python=3.9
-
-    source "${CONDA_ACTIVATE}" ${SL3A_ENV_NAME_TMP}
-
-    "${CONDA}" install -y pytorch=1.13.0 pytorch-cuda=11.6 numpy==1.21.2 -c pytorch -c nvidia
-    "${CONDA}" install -y -c fvcore -c iopath -c conda-forge fvcore=0.1.5.post20221221 iopath=0.1.9 ffmpeg=4.3 cmake=3.26.4
-    "${CONDA}" install -y pytorch3d=0.7.4 -c pytorch3d
-
-    pip install -r "${SELF_DIR}/requirements.txt"
-
-    exit 0
+if [ ! -f "${CONDA}" -o ! -f "${CONDA_ACTIVATE}" ]; then
+    echo "\"conda\" not found, check setup_conda_bin.sh"
+    exit 255
 fi
 
-if [ ! -z "${SL3A_CONDA_FINALIZE_SETUP}" ]; then
-    if "${CONDA}" env list | grep -qw "^${SL3A_ENV_NAME}"; then
-        exit 0
-    fi
-
-    "${CONDA}" rename -n "${SL3A_ENV_NAME_TMP}" "${SL3A_ENV_NAME}"
-
-    exit 0
+if "${CONDA}" env list | grep -qw "^${SL3A_ENV_NAME}"; then
+    "${CONDA}" env remove -n ${SL3A_ENV_NAME}
 fi
 
-# 1. conda renaming is required to ensure consistency of the new environment with respect to the many steps involved in its setup
-# 2. conda renaming cannot be done from within the activated environment, and deactivation is unavailable in unattended mode, so
-#    we just split installation into two executions of this very script activating two different code paths.
-SL3A_CONDA_PERFORM_SETUP=1 bash "$0" "$@"
-SL3A_CONDA_FINALIZE_SETUP=1 bash "$0" "$@"
+"${CONDA}" install -y -n base --solver classic conda-libmamba-solver
+"${CONDA}" config --set solver libmamba
 
+"${CONDA}" create -y -n ${SL3A_ENV_NAME} python=3.9
+
+source "${CONDA_ACTIVATE}" ${SL3A_ENV_NAME}
+
+"${CONDA}" install -y pytorch=1.13.0 pytorch-cuda=11.6 numpy==1.21.2 -c pytorch -c nvidia
+"${CONDA}" install -y -c fvcore -c iopath -c conda-forge fvcore=0.1.5.post20221221 iopath=0.1.9 ffmpeg=4.3 cmake=3.26.4
+"${CONDA}" install -y pytorch3d=0.7.4 -c pytorch3d
+pip install -r "${SELF_DIR}/requirements.txt"
+
+touch "${SL3A_CONDA_ROOT}/.marker.env.completed"
 echo "Conda environment \"${SL3A_ENV_NAME}\" is now installed"
